Move handling of destroyed folders into 'history' object.
If a thunar window is visiting a folder which gets destoryed (rmdir),
it will visit a parent, or $HOME.
However it currently leaves the destroyed directory on the history
so 'back' will go to it.  This feels clumsy.

So move the handling of destroyed directories into thunar-history
where it is easier to get the history right.

Also use the one piece of code for both
  - folder disappears
  - pre-unmount (folder is about to disappear).

This leads to more consistent behaviour.

Signed-off-by: NeilBrown <neilb@suse.de>