thunar-history - guard against non-existent folders in history.
It is quite possible for a non-existent folder to be in the
history - e.g. it is was removed when not currently visited.

If 'forward' or 'back' find this folder they will allow you do
visit it, which is confusing.

So test for folder not existing, and use the new general handler
for destroyed folders to deal with it.

Signed-off-by: NeilBrown <neilb@suse.de>