Add new option for how to handle destroyed folders.
While a usb stick is inserted and removed repeatedly, each insertion
will add a new thunar window, but removal does not destroy them.
I find this inconvenient.

So add a new mechanism for handling destroyed windows.
Rather than finding a parent, it find a recent entry in the
history.  If there isn't one it closes the window.

So:
 - add a new 'advanced' config option
 - change to thunar_history_current_directory_destroy to check the
   option and optionally move around in the history to an existing
   directory, or fail.
 - change callers from thunar_window to test the return status and
   close the window on 'failure'.

Signed-off-by: NeilBrown <neilb@suse.de>